requirements.txt
streamlit
streamlit-extras
torch
torchvision
opencv-python
numpy
pandas
requests
flask
Pillow